[Refactor] Unified the TabletInfo in exec and exec/vectorized There are two TabletInfos in different directories. I think we can unified it into one file
