+++ be/src/exec/tablet_info.h
@@ -12,26 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// This file is based on code available under the Apache license here:
-//   https://github.com/apache/incubator-doris/blob/master/be/src/exec/tablet_info.h
-
-// Licensed to the Apache Software Foundation (ASF) under one
-// or more contributor license agreements.  See the NOTICE file
-// distributed with this work for additional information
-// regarding copyright ownership.  The ASF licenses this file
-// to you under the Apache License, Version 2.0 (the
-// "License"); you may not use this file except in compliance
-// with the License.  You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
 #pragma once
 
 #include <cstdint>
@@ -39,6 +19,7 @@
 #include <unordered_map>
 #include <vector>
 
+#include "column/column.h"
 #include "common/object_pool.h"
 #include "common/status.h"
 #include "gen_cpp/Descriptors_types.h"
@@ -48,6 +29,7 @@
 namespace starrocks {
 
 class MemPool;
+class RuntimeState;
 
 struct OlapTableIndexSchema {
     int64_t index_id;
@@ -97,16 +79,8 @@ private:
 };
 
 using OlapTableIndexTablets = TOlapTableIndexTablets;
-// struct TOlapTableIndexTablets {
-//     1: required i64 index_id
-//     2: required list<i64> tablets
-// }
 
 using TabletLocation = TTabletLocation;
-// struct TTabletLocation {
-//     1: required i64 tablet_id
-//     2: required list<i64> node_ids
-// }
 
 class OlapTableLocationParam {
 public:
@@ -161,6 +135,99 @@ public:
 
 private:
     std::unordered_map<int64_t, NodeInfo> _nodes;
+};
+
+struct ChunkRow {
+    ChunkRow() = default;
+    ChunkRow(Columns* columns_, uint32_t index_) : columns(columns_), index(index_) {}
+
+    Columns* columns = nullptr;
+    uint32_t index = 0;
+};
+
+struct OlapTablePartition {
+    int64_t id = 0;
+    ChunkRow start_key;
+    ChunkRow end_key;
+    int64_t num_buckets = 0;
+    std::vector<OlapTableIndexTablets> indexes;
+};
+
+struct PartionKeyComparator {
+    // return true if lhs < rhs
+    // 'nullptr' is max value, but 'null' is min value
+    bool operator()(const ChunkRow* lhs, const ChunkRow* rhs) const {
+        if (lhs->columns == nullptr) {
+            return false;
+        } else if (rhs->columns == nullptr) {
+            return true;
+        }
+        DCHECK_EQ(lhs->columns->size(), rhs->columns->size());
+
+        for (size_t i = 0; i < lhs->columns->size(); ++i) {
+            int cmp = (*lhs->columns)[i]->compare_at(lhs->index, rhs->index, *(*rhs->columns)[i], -1);
+            if (cmp != 0) {
+                return cmp < 0;
+            }
+        }
+        // equal, return false
+        return false;
+    }
+};
+
+// store an olap table's tablet information
+class OlapTablePartitionParam {
+public:
+    OlapTablePartitionParam(std::shared_ptr<OlapTableSchemaParam> schema, const TOlapTablePartitionParam& param);
+    ~OlapTablePartitionParam();
+
+    Status init(RuntimeState* state);
+
+    Status prepare(RuntimeState* state);
+    Status open(RuntimeState* state);
+    void close(RuntimeState* state);
+
+    int64_t db_id() const { return _t_param.db_id; }
+    int64_t table_id() const { return _t_param.table_id; }
+    int64_t version() const { return _t_param.version; }
+
+    // `invalid_row_index` stores index that chunk[index]
+    // has been filtered out for not being able to find tablet.
+    // it could be any row, becauset it's just for outputing error message for user to diagnose.
+    Status find_tablets(Chunk* chunk, std::vector<OlapTablePartition*>* partitions, std::vector<uint32_t>* indexes,
+                        std::vector<uint8_t>* selection, int* invalid_row_index);
+
+    const std::vector<OlapTablePartition*>& get_partitions() const { return _partitions; }
+
+    bool is_un_partitioned() const { return _partition_columns.empty(); }
+
+private:
+    Status _create_partition_keys(const std::vector<TExprNode>& t_exprs, ChunkRow* part_key);
+
+    void _compute_hashes(Chunk* chunk, std::vector<uint32_t>* indexes);
+
+    // check if this partition contain this key
+    bool _part_contains(OlapTablePartition* part, ChunkRow* key) const {
+        if (part->start_key.columns == nullptr) {
+            // start_key is nullptr means the lower bound is boundless
+            return true;
+        }
+        return !PartionKeyComparator()(key, &part->start_key);
+    }
+
+private:
+    std::shared_ptr<OlapTableSchemaParam> _schema;
+    TOlapTablePartitionParam _t_param;
+
+    std::vector<SlotDescriptor*> _partition_slot_descs;
+    std::vector<SlotDescriptor*> _distributed_slot_descs;
+    Columns _partition_columns;
+    std::vector<Column*> _distributed_columns;
+    std::vector<ExprContext*> _partitions_expr_ctxs;
+
+    ObjectPool _obj_pool;
+    std::vector<OlapTablePartition*> _partitions;
+    std::map<ChunkRow*, OlapTablePartition*, PartionKeyComparator> _partitions_map;
 };
 
 } // namespace starrocks
